Unit test for removing people and enqueuing people for removal
