+++ PART/Source/CprBroker/PartInterface.Tests/PartInterface.Tests.csproj
@@ -53,6 +53,9 @@
     <Reference Include="CprBroker.Installers.EventBrokerInstallers">
       <HintPath>..\..\EventBroker\Output\CprBroker.Installers.EventBrokerInstallers.dll</HintPath>
     </Reference>
+    <Reference Include="CprBroker.PartInterface.Tracking">
+      <HintPath>..\Web\bin\CprBroker.PartInterface.Tracking.dll</HintPath>
+    </Reference>
     <Reference Include="CprBroker.Schemas">
       <HintPath>..\..\Core\Output\CprBroker.Schemas.dll</HintPath>
     </Reference>
